Create final.py
Finalized version. Mainly reorganizing code.

- added docstrings for revisiting.

- renamed variable from "earnings" to "revenue" and removed from coffeemachine.py to be in main body.

- removed isinstance from main().

- renamed resource_sufficient(drink) to can_we_make_it(menu_ingredients), reworked to return True or False.

- tidied up process_payment() a little.

- check_transaction_success(total, drink) renamed to is_transaction_successful(payment, drink_cost) because it looked nicer and easier to follow along. Adjusted to return True or False.

- combined if/elif of total < drink_cost and total > drink_cost into if payment >= drink_cost to catch all.

- new function make_coffee(drink_name, order_ingredients): moved "elif resources[ingredients] >= quantity:" from resource_sufficient to only subtract when make_coffee is available.

Notes: 
Although ininstance() was fun to use, it was unnecessary. I had to reread the program requirements several times and it took several days to understand that I should have just kept it simple instead of trying to mix-and-match various functions when I'm still learning all of this and just went one segment at a time.